Fix EDS cart issues
So I got a successful result from putting items in the cart from the
discovery tab and then having it being sent through (only books) and
with all the information showing.

No success from the Koha tab - which is what mainly our clients will be
searching from for just books and things within our physical
collections. When trying to send a cart with only books it just spins
around for a while and then the error page turns up. (The message, this
webpage cannot be displayed). So this really in our books isn’t working
at all from the Koha tab side of things.
I also have a bug that needs registering with you about books staying in
the cart and not leaving, even when deleting them out of there and
logging back in and out.