EDS plugin filling Koha logs with needless warnings #53
For the most part, these are warnings about the experimental use
of given/when.

Koha doesn't use given/when for this reason, and there have been
patches to remove its use in the past:
https://bugs.koha-community.org/bugzilla3/show_bug.cgi?id=11478
https://bugs.koha-community.org/bugzilla3/show_bug.cgi?id=11479

I thought there was a line in Koha's coding guidelines but I
haven't found it. It might be one of those things which is
implicitly understood in the Koha community but isn't recorded
explicitly anywhere.

This patch removes the use of given/when in eds-methods.pl and
replaces it with if statements.