Allow Koha to determine the real IP address of the request
Koha has a middleware to determine the real originating IP of a request
by checking trusted proxy ip address, X-Forwarded-For and with bug
39789, arbitrary other headers.

The EDS plugin should call C4::Context::set_remote_address For the EDS to function in some
circumstances where X-Forwarded-For is insufficient. This also removes the
need to check X-Forwarded-For specifically as Koha already handles that.